+++ frontend/src/Components/RoadDetails/ConditionsGraph.tsx
