+++ frontend/src/Components/RoadDetails/RoadImage.tsx
@@ -3,7 +3,6 @@ import React from 'react';
 const maproadStyles = {
   margin: 0,
   padding: 20,
-  color: '#333',
 };
 
 const road_area = <h1 style={maproadStyles}>Road surface image</h1>;
