--- frontend/src/css/road_conditions.css
@@ -1,13 +1,0 @@
-.road-conditions-page-wrapper {
-  height: calc(100vh - var(--navbar-height));
-  display: flex;
-  flex-direction: column;
-}
-
-.road-conditions-graph {
-  position: relative;
-  width: 100vw;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
